refactor(EG-1040): omics: refactor out EGWorkflowStepper
