+++ packages/front-end/src/app/pages/labs/[labId]/run-workflow/[workflowId].vue
@@ -84,6 +84,8 @@
    * Reads the workflow details, schema, and parameters from the API and initializes the pipeline run store
    */
   async function initializeWorkflowData() {
+    uiStore.setRequestPending('loadOmicsWorkflow');
+
     runStore.updateWipOmicsRun(omicsRunTempId.value, {
       laboratoryId: labId!,
       workflowId: workflowId!,
@@ -113,6 +115,8 @@
         paramsRequired: paramsRequired,
       });
     }
+
+    uiStore.setRequestComplete('loadOmicsWorkflow');
   }
 
   /**
@@ -130,6 +134,121 @@
       initializeWorkflowData();
       resetStepperKey.value++;
     }, 100);
+  }
+
+  // Note: the UTabs :ui attribute has to be defined locally in this file - if it is imported from another file,
+  //  Tailwind won't pick up and include the classes used and styles will be missing.
+  // To keep the tab styling consistent throughout the app, any changes made here need to be duplicated to all other
+  //  UTabs that use an "EGTabsStyles" as input to the :ui attribute.
+  const EGTabsStyles = {
+    base: 'focus:outline-none',
+    list: {
+      base: '!flex rounded-none mb-6 mt-0',
+      padding: 'p-0',
+      height: 'h-14',
+      marker: {
+        background: '',
+        shadow: '',
+      },
+      tab: {
+        base: 'font-serif w-auto mr-3 rounded-xl border border-solid',
+        background: '',
+        active: 'text-white bg-primary border-primary',
+        inactive: 'font-serif text-text-body border-background-dark-grey',
+        height: '',
+        padding: 'px-5 py-2',
+        size: 'text-sm',
+      },
+    },
+  };
+
+  // stepper data handling
+
+  const selectedStepIndex = ref(0);
+
+  const steps = ref([
+    {
+      disabled: false,
+      key: 'details',
+      label: 'Run Details',
+    },
+    {
+      disabled: true,
+      key: 'upload',
+      label: 'Upload Data',
+    },
+    {
+      disabled: true,
+      key: 'parameters',
+      label: 'Edit Parameters',
+    },
+    {
+      disabled: true,
+      key: 'review',
+      label: 'Review Pipeline',
+    },
+  ]);
+
+  /**
+   * Set the enabled state of a step in the stepper
+   * @param step
+   * @param isEnabled
+   */
+  function setStepEnabled(stepKey: string, isEnabled: boolean) {
+    const stepIndex = steps.value.findIndex((step) => step.key === stepKey);
+
+    if (stepIndex === -1) throw new Error(`no step found with key "${stepKey}"`);
+
+    if (isEnabled) {
+      steps.value[stepIndex].disabled = false;
+    } else {
+      // If the step is disabled, disable all subsequent steps
+      disableStepsFrom(stepIndex);
+    }
+  }
+
+  /**
+   * Disable all steps from the given index
+   * @param index
+   */
+  function disableStepsFrom(index: number) {
+    for (let i = index; i < steps.value.length; i++) {
+      steps.value[i].disabled = true;
+    }
+  }
+
+  function nextStep(val: string) {
+    setStepEnabled(val, true);
+    selectedStepIndex.value = clampIndex(selectedStepIndex.value + 1);
+  }
+
+  function clampIndex(index: number) {
+    return Math.min(steps.value.length - 1, Math.max(0, index));
+  }
+
+  function previousStep() {
+    selectedStepIndex.value = clampIndex(selectedStepIndex.value - 1);
+  }
+
+  function disableAllSteps() {
+    steps.value.forEach((step) => (step.disabled = true));
+  }
+
+  function enableAllSteps() {
+    steps.value.forEach((step) => (step.disabled = false));
+  }
+
+  function handleSubmitLaunchRequest() {
+    disableAllSteps();
+  }
+
+  function handleSubmitLaunchRequestError() {
+    enableAllSteps();
+  }
+
+  function handleLaunchSuccess() {
+    hasLaunched.value = true;
+    selectedStepIndex.value = -1;
   }
 </script>
 
@@ -150,15 +269,97 @@
   </template>
 
   <template v-else>
-    <EGRunWorkflowStepper
-      @has-launched="hasLaunched = true"
-      :schema="schema"
-      :params="wipOmicsRun?.params"
-      @reset-run-pipeline="resetRunPipeline()"
-      :key="resetStepperKey"
-      :workflow-id="workflowId"
-      :lab-name="labName"
-    />
+    <UTabs :items="steps" :ui="EGTabsStyles" v-model="selectedStepIndex" :key="selectedStepIndex">
+      <!-- tab rendering -->
+      <template #default="{ item, index, selected }">
+        <div class="relative flex items-center gap-2 truncate">
+          <UIcon
+            v-if="selectedStepIndex > index || hasLaunched"
+            name="i-heroicons-check-20-solid"
+            class="text-primary h-4 w-4 flex-shrink-0"
+          />
+          <span :class="selectedStepIndex > index || hasLaunched ? 'text-primary' : ''">{{ item.label }}</span>
+          <span v-if="selected" class="bg-primary-500 dark:bg-primary-400 absolute -right-4 h-2 w-2 rounded-full" />
+        </div>
+      </template>
+
+      <!-- step rendering -->
+      <template #item="{ item, index }">
+        <div v-if="!hasLaunched">
+          <!-- Run Details -->
+          <template v-if="steps[selectedStepIndex].key === 'details'">
+            <EGRunFormRunDetails
+              pipeline-or-workflow="Workflow"
+              :pipeline-or-workflow-name="workflow?.name"
+              :initial-run-name="wipOmicsRun?.runName || ''"
+              :pipeline-or-workflow-description="workflow?.description || ''"
+              :wip-run-update-function="runStore.updateWipOmicsRun"
+              :wip-run-temp-id="omicsRunTempId"
+              @next-step="() => nextStep('upload')"
+              @step-validated="($event) => setStepEnabled('upload', $event)"
+            />
+          </template>
+
+          <!-- Upload Data -->
+          <template v-if="steps[selectedStepIndex].key === 'upload'">
+            <EGRunFormUploadData
+              :lab-id="labId"
+              :lab-name="labName"
+              :pipeline-or-workflow-name="workflow.name"
+              :wip-run="wipOmicsRun"
+              :wip-run-update-function="runStore.updateWipOmicsRun"
+              :wip-run-temp-id="omicsRunTempId"
+              @next-step="() => nextStep('parameters')"
+              @previous-step="() => previousStep()"
+              @step-validated="($event) => setStepEnabled('parameters', $event)"
+            />
+          </template>
+
+          <!-- Edit Parameters -->
+          <template v-if="steps[selectedStepIndex].key === 'parameters'">
+            <EGRunWorkflowFormEditParameters
+              :params="wipOmicsRun?.params"
+              :schema="schema"
+              :omics-run-temp-id="omicsRunTempId"
+              @next-step="() => nextStep('review')"
+              @previous-step="() => previousStep()"
+            />
+          </template>
+
+          <!-- Review Pipeline -->
+          <template v-if="steps[selectedStepIndex].key === 'review'">
+            <EGRunWorkflowFormReview
+              :schema="schema"
+              :params="wipOmicsRun?.params"
+              :lab-id="labId"
+              :omics-run-temp-id="omicsRunTempId"
+              :s3-bucket="wipOmicsRun?.s3Bucket"
+              :s3-path="wipOmicsRun?.s3Path"
+              :run-name="wipOmicsRun?.runName"
+              :transaction-id="wipOmicsRun?.transactionId"
+              :workflow-id="workflowId"
+              :workflow-name="workflow.name"
+              @submit-launch-request="() => handleSubmitLaunchRequest()"
+              @submit-launch-request-error="() => handleSubmitLaunchRequestError()"
+              @has-launched="() => handleLaunchSuccess()"
+              @previous-tab="() => previousStep()"
+            />
+          </template>
+        </div>
+      </template>
+    </UTabs>
+
+    <!-- post-launch rendering -->
+    <template v-if="hasLaunched">
+      <EGEmptyDataCTA
+        message="Your Workflow Run has Launched! Check on your progress via Runs."
+        :primary-button-action="() => $router.push(`/labs/${labId}?tab=Lab+Runs`)"
+        primary-button-label="Back to Runs"
+        :secondary-button-action="() => resetRunPipeline()"
+        secondary-button-label="Launch Another Workflow Run"
+        img-src="/images/empty-state-launched.jpg"
+      />
+    </template>
 
     <EGDialog
       action-label="Cancel Workflow Run"
